requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
envparse==0.2.0
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.22.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
zipp==3.8.0
